+++ tests/run.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
-flake8 . --exclude=build,dist
+flake8 . --exclude=build,dist --extend-ignore=E251
 py.test "$@"